feat: implementar validações de integridade para criação, atualização e remoção de usuários

+++ apps/teachers-service/src/app.controller.ts
@@ -2,6 +2,7 @@ import { Controller } from '@nestjs/common';
 import { AppService } from './app.service';
 import { EventPattern, MessagePattern, Payload } from '@nestjs/microservices';
 import { TeacherUserDto } from '@obg/schemas';
+import { UserRole } from '@obg/enums';
 
 @Controller()
 export class AppController {
@@ -9,12 +10,20 @@ export class AppController {
 
   @EventPattern('user.created')
   async create(@Payload() data: TeacherUserDto) {
+    if (data.role !== UserRole.TEACHER) {
+      return; // Ignora se o role não for TEACHER
+    }
     return await this.appService.create(data);
   }
 
   @MessagePattern('teacher.findAll')
   async findAll(@Payload() payload: { page: number; pageSize: number }) {
     return await this.appService.findAll(payload.page, payload.pageSize);
+  }
+
+  @MessagePattern('teacher.findOneByEmail')
+  async findOneByEmail(@Payload() email: string) {
+    return await this.appService.findOneByEmail(email);
   }
 
   @MessagePattern('teacher.findOne')
@@ -24,11 +33,17 @@ export class AppController {
 
   @EventPattern('user.updated')
   async update(@Payload() data: TeacherUserDto) {
+    if (data.role !== UserRole.TEACHER) {
+      return; // Ignora se o role não for TEACHER
+    }
     return await this.appService.update(data);
   }
 
   @EventPattern('user.deleted')
   async remove(@Payload() data: TeacherUserDto) {
+    if (data.role !== UserRole.TEACHER) {
+      return; // Ignora se o role não for TEACHER
+    }
     return await this.appService.remove(data);
   }
 }
